Rename core table construction module
